feat: add state machine for simulation

+++ src/main.rs
@@ -5,9 +5,9 @@ mod simulation;
 
 fn main() -> std::io::Result<()> {
     let target_str: &str = "hello world";
-    let population_size: u32 = 48;
-    let num_fit_to_keep: u32 = 5;
-    let num_columns: u32 = 4;
+    let population_size: usize = 48;
+    let num_fit_to_keep: usize = 5;
+    let num_columns: usize = 4;
     let mutation_prob: f32 = 0.15;
 
     let mut simulation = Simulation::new(
